refactor: 폴더명을 constants에서 constant로 변경
